requirements.txt
# Core FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# PDF Processing
PyMuPDF==1.23.21
pdfplumber==0.10.3

# NLP & Embeddings
numpy<2.0.0
sentence-transformers==2.3.1
transformers==4.36.2

# Vector Stores
faiss-cpu==1.7.4
qdrant-client==1.7.1

# LLM & API clients
httpx==0.26.0
openai==1.12.0

# Graph & compatibility
networkx==3.2.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
